pyproject.toml
[tool.poetry]
name = "drf-test-project"
version = "0.1.0"
description = "Test project to learn how-to Django DRF"
authors = ["AlanZhuman"]
license = "None"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
annotated-types = "0.7.0"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
djangorestframework = "3.15.2"
drf-yasg = "^1.21.7"
idna = "^3.10"
importlib-resources = "6.4.5"
inflection = "0.5.1"
jsonschema = "4.23.0"
packaging = "24.1"
python-dotenv = "1.0.1"
pytz = "2024.2"
pyyaml = "6.0.2"
referencing = "0.35.1"
swagger-spec-validator = "3.0.4"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
rpds-py = "0.20.0"
setuptools = "75.1.0"
sqlparse = "0.5.1"
celery = "^5.4.0"
redis = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
